moving mocks to mock_gcloud
